[Refactor] Restructure "net" module: add separate folder
